`getSafeWithOwners` using types (#796)
Fixes #780 

This PR uses the typescript `types` to pass the parameters to
`getSafeWithOwners(...)` as objects. All the parameters are kept as
`readonly` because it is not supposed to change within the function
(though `optional` parameters are assigned default values during
restructuring).

Bonus Point: Found that `getSafeWithSingleton(...)` didn't use the
`threshold` and `fallbackHandler` parameters in the current codebase,
thus removing them for now (please comment if it should be reverted).